dashboard/requirements.txt
kafka-python
matplotlib
flask
flask-cors